Package shuffling, docs
