+++ core/js/src/main/scala/cats/effect/ThunkImpl.scala
@@ -14,13 +14,8 @@
  * limitations under the License.
  */
 
-package cats.effect.util
+package cats.effect
 
-private abstract class Thunk {
-  def asFunction0[A](thunk: => A): Function0[A]
-}
-
-private[effect] object Thunk {
-  private[this] val impl: Thunk = new ThunkImpl
-  def asFunction0[A](thunk: => A): Function0[A] = impl.asFunction0(thunk)
+private class ThunkImpl extends Thunk {
+  override def asFunction0[A](thunk: => A) = () => thunk
 }
